+++ src/controllers/Social.ts
@@ -1,14 +1,13 @@
 import type { AccessKeyView } from '@near-js/types';
 import BigNumber from 'bignumber.js';
-import {
-  type Account,
-  type Connection,
-  transactions,
-  utils,
-} from 'near-api-js';
+import { providers, transactions, utils } from 'near-api-js';
 
 // constants
-import { GAS_FEE_IN_ATOMIC_UNITS, ONE_YOCTO } from '@app/constants';
+import {
+  GAS_FEE_IN_ATOMIC_UNITS,
+  networkRPCs,
+  ONE_YOCTO,
+} from '@app/constants';
 
 // enums
 import { ChangeMethodEnum, ViewMethodEnum } from '@app/enums';
@@ -18,17 +17,18 @@ import {
   AccountNotFoundError,
   InvalidAccountIdError,
   KeyNotAllowedError,
+  UnknownNetworkError,
 } from '@app/errors';
 
 // types
 import type {
   IGetOptions,
-  IGetVersionOptions,
   IGrantWritePermissionWithAccountIdOptions,
   IGrantWritePermissionWithPublicKeyOptions,
   IIsWritePermissionGrantedWithAccountIdOptions,
   IIsWritePermissionGrantedWithPublicKeyOptions,
   INewSocialOptions,
+  IRPCOptions,
   ISetOptions,
   IStorageDepositOptions,
   IStorageWithdrawOptions,
@@ -37,27 +37,79 @@ import type {
   ISocialDBContractGrantWritePermissionArgs,
   ISocialDBContractSetArgs,
   ISocialDBContractStorageBalance,
-  IStorageBalanceOfOptions,
   ISocialDBContractIsWritePermissionGrantedArgs,
   ISocialDBContractStorageWithdrawArgs,
   ISocialDBContractStorageDepositArgs,
   IKeysOptions,
   ISocialApiServerKeysArgs,
   ISocialDBContractKeysArgs,
+  IAccount,
 } from '@app/types';
 
 // utils
 import calculateRequiredDeposit from '@app/utils/calculateRequiredDeposit';
 import parseKeysFromData from '@app/utils/parseKeysFromData';
+import rpcURLFromNetworkID from '@app/utils/rpcURLFromNetworkID';
 import validateAccountId from '@app/utils/validateAccountId';
+import viewAccessKeyList from '@app/utils/rpcQueries/viewAccessKeyList';
+import viewFunction from '@app/utils/rpcQueries/viewFunction';
 
 export default class Social {
-  private contractId: string;
+  // private variables
+  private readonly _contractId: string;
+  private readonly _provider: providers.JsonRpcProvider;
   private apiServer: string;
 
   constructor(options?: INewSocialOptions) {
-    this.contractId = options?.contractId || 'social.near';
+    this._contractId = options?.contractId || 'social.near';
+    this._provider = Social._initializeProvider(options?.network);
     this.apiServer = options?.apiServer || 'https://api.near.social';
+  }
+
+  /**
+   * private static methods
+   */
+
+  /**
+   * Initializes the provider with the supplied network options. If the network options are empty, the default mainnet
+   * is used.
+   * @param {string | IRPCOptions} networkIDOrRPCOptions - [optional] a network ID or the RPC options to initialize a
+   * provider.
+   * @returns {providers.JsonRpcProvider} an initialized provider to query the network with.
+   * @throws {UnknownNetworkError} if a network ID is supplied, but is not known.
+   * @private
+   * @static
+   */
+  private static _initializeProvider(
+    networkIDOrRPCOptions?: string | IRPCOptions
+  ): providers.JsonRpcProvider {
+    let url: string | null;
+
+    // if there is no network id/rpc details, default to mainnet
+    if (!networkIDOrRPCOptions) {
+      return new providers.JsonRpcProvider({ url: networkRPCs.mainnet });
+    }
+
+    // if there is a network id, attempt to get the rpc url
+    if (typeof networkIDOrRPCOptions === 'string') {
+      url = rpcURLFromNetworkID(networkIDOrRPCOptions);
+
+      if (!url) {
+        throw new UnknownNetworkError(networkIDOrRPCOptions);
+      }
+
+      return new providers.JsonRpcProvider({ url });
+    }
+
+    // otherwise, use the rpc details
+    return new providers.JsonRpcProvider({
+      url: networkIDOrRPCOptions.url,
+      ...(networkIDOrRPCOptions.apiKey && {
+        headers: {
+          ['X-Api-Key']: networkIDOrRPCOptions.apiKey,
+        },
+      }),
+    });
   }
 
   /**
@@ -66,17 +118,19 @@ export default class Social {
 
   /**
    * Gets the access key view.
-   * @param {Account} account - an initialized account.
-   * @param {string | utils.PublicKey} publicKey - the public key of the access key to query.
+   * @param {IAccount} account - the account ID and public key of the account.
    * @returns {Promise<AccessKeyView | null>} a promise that resolves to the access key view or null if the access key
    * for the given public key does not exist.
    * @private
    */
-  private async _accessKeyView(
-    account: Account,
-    publicKey: string | utils.PublicKey
-  ): Promise<AccessKeyView | null> {
-    const accessKeys = await account.getAccessKeys();
+  private async _accessKeyView({
+    accountID,
+    publicKey,
+  }: IAccount): Promise<AccessKeyView | null> {
+    const accessKeys = await viewAccessKeyList({
+      accountID,
+      provider: this._provider,
+    });
 
     return (
       accessKeys.find((value) => value.public_key === publicKey.toString())
@@ -86,27 +140,47 @@ export default class Social {
 
   /**
    * Queries the node to get the latest block hash.
-   * @param {Connection} connection - an initialized NEAR connection.
    * @returns {Promise<string>} a promise that resolves to the latest block hash. The hash will be a base58 encoded string.
    * @private
    */
-  private async _latestBlockHash(connection: Connection): Promise<string> {
-    const { sync_info } = await connection.provider.status();
+  private async _latestBlockHash(): Promise<string> {
+    const { sync_info } = await this._provider.status();
 
     return sync_info.latest_block_hash;
   }
 
-  private async _storageBalanceOf({
-    accountId,
-    signer,
-  }: IStorageBalanceOfOptions): Promise<ISocialDBContractStorageBalance | null> {
-    return await signer.viewFunction({
+  private async _storageBalanceOf(
+    accountID: string
+  ): Promise<ISocialDBContractStorageBalance | null> {
+    const result = await viewFunction({
       args: {
-        account_id: accountId,
+        account_id: accountID,
       },
-      contractId: this.contractId,
-      methodName: ViewMethodEnum.StorageBalanceOf,
+      contractId: this._contractId,
+      method: ViewMethodEnum.StorageBalanceOf,
+      provider: this._provider,
     });
+
+    if (this._isStorageBalance(result)) {
+      return result;
+    } else if (result === null) {
+      return null;
+    } else {
+      throw new Error('Unexpected response format from storage_balance_of');
+    }
+  }
+
+  private _isStorageBalance(
+    data: unknown
+  ): data is ISocialDBContractStorageBalance {
+    return (
+      typeof data === 'object' &&
+      data !== null &&
+      'total' in data &&
+      'available' in data &&
+      typeof (data as ISocialDBContractStorageBalance).total === 'string' &&
+      typeof (data as ISocialDBContractStorageBalance).available === 'string'
+    );
   }
 
   private _uniqueAccountIdsFromKeys(keys: string[]): string[] {
@@ -125,11 +199,11 @@ export default class Social {
 
   /**
    * Reads the data for given set of keys.
-   * @param {IGetOptions} options - the signer and a set of keys to read.
+   * @param {IGetOptions} options - the set of keys to read and other options.
    * @returns {Promise<Record<string, unknown>>} a promise that resolves to the given data.
+   * @public
    */
   public async get({
-    signer,
     keys,
     blockHeight,
     returnDeleted,
@@ -159,20 +233,23 @@ export default class Social {
         })
       ).json();
     } else {
-      return await signer?.viewFunction({
-        contractId: this.contractId,
-        methodName: ViewMethodEnum.Get,
-        args: {
-          keys,
-          ...((returnDeleted || withBlockHeight || withNodeId) && {
-            options: {
-              with_block_height: withBlockHeight,
-              with_node_id: withNodeId,
-              return_deleted: returnDeleted,
-            },
-          }),
-        } as ISocialDBContractGetArgs,
-      });
+      const args: ISocialDBContractGetArgs = {
+        keys,
+        ...((returnDeleted || withBlockHeight || withNodeId) && {
+          options: {
+            with_block_height: withBlockHeight,
+            with_node_id: withNodeId,
+            return_deleted: returnDeleted,
+          },
+        }),
+      };
+
+      return (await viewFunction({
+        args,
+        contractId: this._contractId,
+        method: ViewMethodEnum.Get,
+        provider: this._provider,
+      })) as Record<string, unknown>;
     }
   }
 
@@ -232,12 +309,21 @@ export default class Social {
   /**
    * Gets the current version of the social contract.
    * @returns {Promise<string>} a promise that resolves to the current version of the contract.
+   * @public
    */
-  public async getVersion({ signer }: IGetVersionOptions): Promise<string> {
-    return await signer.viewFunction({
-      contractId: this.contractId,
-      methodName: ViewMethodEnum.GetVersion,
+  public async getVersion(): Promise<string> {
+    const version = await viewFunction({
+      contractId: this._contractId,
+      method: ViewMethodEnum.GetVersion,
+      provider: this._provider,
     });
+
+    if (typeof version !== 'string') {
+      throw new Error(
+        `Unexpected response format from get_version: ${JSON.stringify(version)}`
+      );
+    }
+    return version;
   }
 
   /**
@@ -248,13 +334,14 @@ export default class Social {
    * and sent to the network.
    * @throws {InvalidAccountIdError} if the grantee account ID or the account ID specified in the keys is invalid.
    * @throws {KeyNotAllowedError} if account IDs specified in the keys does not match the signer (granter) account ID.
+   * @public
    */
   public async grantWritePermission(
     options:
       | IGrantWritePermissionWithAccountIdOptions
       | IGrantWritePermissionWithPublicKeyOptions
   ): Promise<transactions.Transaction> {
-    const { blockHash, keys, nonce, publicKey, signer } = options;
+    const { account, blockHash, keys, nonce } = options;
     let accessKeyView: AccessKeyView | null;
     let _blockHash: string | null = blockHash || null;
     let _nonce: bigint | null = nonce || null;
@@ -280,25 +367,25 @@ export default class Social {
       }
 
       // if the key does not belong to the signer (granter) it cannot give grant permission
-      if (accountId !== signer.accountId) {
+      if (accountId !== account.accountID) {
         throw new KeyNotAllowedError(
           value,
-          `key "${value}" does not belong to granter "${signer.accountId}"`
+          `key "${value}" does not belong to granter "${account.accountID}"`
         );
       }
     });
 
     if (!_blockHash) {
-      _blockHash = await this._latestBlockHash(signer.connection);
+      _blockHash = await this._latestBlockHash();
     }
 
     if (!_nonce) {
-      accessKeyView = await this._accessKeyView(signer, publicKey);
+      accessKeyView = await this._accessKeyView(account);
 
       if (!accessKeyView) {
         throw new AccountNotFoundError(
-          signer.accountId,
-          `failed to get nonce for access key for "${signer.accountId}" with public key "${publicKey.toString()}"`
+          account.accountID,
+          `failed to get nonce for access key for "${account.accountID}" with public key "${account.publicKey.toString()}"`
         );
       }
 
@@ -306,9 +393,9 @@ export default class Social {
     }
 
     return transactions.createTransaction(
-      signer.accountId,
-      utils.PublicKey.fromString(publicKey.toString()),
-      this.contractId,
+      account.accountID,
+      utils.PublicKey.fromString(account.publicKey.toString()),
+      this._contractId,
       _nonce,
       [
         transactions.functionCall(
@@ -343,13 +430,14 @@ export default class Social {
    * @returns {Promise<boolean>} a promise that resolves to true, if the grantee account ID has write access for the
    * given key, or false.
    * @throws {InvalidAccountIdError} if the grantee account ID is not a valid account ID.
+   * @public
    */
   public async isWritePermissionGranted(
     options:
       | IIsWritePermissionGrantedWithAccountIdOptions
       | IIsWritePermissionGrantedWithPublicKeyOptions
   ): Promise<boolean> {
-    const { key, signer } = options;
+    const { key } = options;
 
     if (
       (options as IIsWritePermissionGrantedWithAccountIdOptions)
@@ -378,9 +466,7 @@ export default class Social {
       }
     }
 
-    return await signer.viewFunction({
-      contractId: this.contractId,
-      methodName: ViewMethodEnum.IsWritePermissionGranted,
+    const result = await viewFunction({
       args: {
         key,
         ...((options as IIsWritePermissionGrantedWithAccountIdOptions)
@@ -396,7 +482,18 @@ export default class Social {
           ).granteePublicKey.toString(),
         }),
       } as ISocialDBContractIsWritePermissionGrantedArgs,
+      contractId: this._contractId,
+      method: ViewMethodEnum.IsWritePermissionGranted,
+      provider: this._provider,
     });
+
+    if (typeof result !== 'boolean') {
+      throw new Error(
+        `Unexpected response format from isWritePermissionGranted: ${JSON.stringify(result)}`
+      );
+    }
+
+    return result;
   }
 
   /**
@@ -405,33 +502,33 @@ export default class Social {
    * @param {ISetOptions} options - the necessary options to set some data.
    * @returns {Promise<transactions.Transaction>} a promise that resolves to a transaction that is ready to be signed
    * and sent to the network.
+   * @public
    */
   public async set({
+    account,
     blockHash,
     data,
     nonce,
-    publicKey,
     refundUnusedDeposit,
-    signer,
   }: ISetOptions): Promise<transactions.Transaction> {
     const keys = parseKeysFromData(data);
-    let _blockHash: string | null = blockHash || null;
-    let _nonce: bigint | null = nonce || null;
+    let _blockHash = blockHash || null;
+    let _nonce = nonce || null;
     let accessKeyView: AccessKeyView | null;
     let deposit: BigNumber = new BigNumber('1');
     let storageBalance: ISocialDBContractStorageBalance | null;
 
     if (!_blockHash) {
-      _blockHash = await this._latestBlockHash(signer.connection);
+      _blockHash = await this._latestBlockHash();
     }
 
     if (!_nonce) {
-      accessKeyView = await this._accessKeyView(signer, publicKey);
+      accessKeyView = await this._accessKeyView(account);
 
       if (!accessKeyView) {
         throw new AccountNotFoundError(
-          signer.accountId,
-          `failed to get nonce for access key for "${signer.accountId}" with public key "${publicKey.toString()}"`
+          account.accountID,
+          `failed to get nonce for access key for "${account.accountID}" with public key "${account.publicKey.toString()}"`
         );
       }
 
@@ -441,11 +538,10 @@ export default class Social {
     // for each key, check if the signer has been granted write permission for the key
     for (let i = 0; i < keys.length; i++) {
       if (
-        (keys[i].split('/')[0] || '') !== signer.accountId &&
+        (keys[i].split('/')[0] || '') !== account.accountID &&
         !(await this.isWritePermissionGranted({
-          granteePublicKey: publicKey,
+          granteePublicKey: account.publicKey,
           key: keys[i],
-          signer,
         }))
       ) {
         throw new KeyNotAllowedError(
@@ -458,13 +554,10 @@ export default class Social {
     // if the signer is updating their own data, calculate storage deposit
     if (
       this._uniqueAccountIdsFromKeys(keys).find(
-        (value) => value === signer.accountId
+        (value) => value === account.accountID
       )
     ) {
-      storageBalance = await this._storageBalanceOf({
-        accountId: signer.accountId,
-        signer,
-      });
+      storageBalance = await this._storageBalanceOf(account.accountID);
 
       deposit = calculateRequiredDeposit({
         data,
@@ -473,9 +566,9 @@ export default class Social {
     }
 
     return transactions.createTransaction(
-      signer.accountId,
-      utils.PublicKey.fromString(publicKey.toString()),
-      this.contractId,
+      account.accountID,
+      utils.PublicKey.fromString(account.publicKey.toString()),
+      this._contractId,
       _nonce,
       [
         transactions.functionCall(
@@ -502,12 +595,12 @@ export default class Social {
    * @param {IStorageDepositOptions} options - the necessary options to deposit NEAR for covering storage for the account_id or the signer.
    * @returns {Promise<transactions.Transaction>} a promise that resolves to a transaction that is ready to be signed
    * and sent to the network.
+   * @public
    */
   public async storageDeposit({
+    account,
     blockHash,
     nonce,
-    publicKey,
-    signer,
     registrationOnly,
     accountId,
     deposit,
@@ -532,16 +625,16 @@ export default class Social {
     let accessKeyView: AccessKeyView | null;
 
     if (!_blockHash) {
-      _blockHash = await this._latestBlockHash(signer.connection);
+      _blockHash = await this._latestBlockHash();
     }
 
     if (!_nonce) {
-      accessKeyView = await this._accessKeyView(signer, publicKey);
+      accessKeyView = await this._accessKeyView(account);
 
       if (!accessKeyView) {
         throw new AccountNotFoundError(
-          signer.accountId,
-          `failed to get nonce for access key for "${signer.accountId}" with public key "${publicKey.toString()}"`
+          account.accountID,
+          `failed to get nonce for access key for "${account.accountID}" with public key "${account.publicKey.toString()}"`
         );
       }
 
@@ -549,9 +642,9 @@ export default class Social {
     }
 
     return transactions.createTransaction(
-      signer.accountId,
-      utils.PublicKey.fromString(publicKey.toString()),
-      this.contractId,
+      account.accountID,
+      utils.PublicKey.fromString(account.publicKey.toString()),
+      this._contractId,
       _nonce,
       actions,
       utils.serialize.base_decode(_blockHash)
@@ -563,13 +656,13 @@ export default class Social {
    * @param {IStorageWithdrawOptions} options - define the amount to be withdrawn.
    * @returns {Promise<transactions.Transaction>} a promise that resolves to a transaction that is ready to be signed
    * and sent to the network.
+   * @public
    */
   public async storageWithdraw({
+    account,
     blockHash,
     amount,
     nonce,
-    publicKey,
-    signer,
   }: IStorageWithdrawOptions): Promise<transactions.Transaction> {
     const actions: transactions.Action[] = [];
 
@@ -590,16 +683,16 @@ export default class Social {
     let accessKeyView: AccessKeyView | null;
 
     if (!_blockHash) {
-      _blockHash = await this._latestBlockHash(signer.connection);
+      _blockHash = await this._latestBlockHash();
     }
 
     if (!_nonce) {
-      accessKeyView = await this._accessKeyView(signer, publicKey);
+      accessKeyView = await this._accessKeyView(account);
 
       if (!accessKeyView) {
         throw new AccountNotFoundError(
-          signer.accountId,
-          `failed to get nonce for access key for "${signer.accountId}" with public key "${publicKey.toString()}"`
+          account.accountID,
+          `failed to get nonce for access key for "${account.accountID}" with public key "${account.publicKey.toString()}"`
         );
       }
 
@@ -607,19 +700,12 @@ export default class Social {
     }
 
     return transactions.createTransaction(
-      signer.accountId,
-      utils.PublicKey.fromString(publicKey.toString()),
-      this.contractId,
+      account.accountID,
+      utils.PublicKey.fromString(account.publicKey.toString()),
+      this._contractId,
       _nonce,
       actions,
       utils.serialize.base_decode(_blockHash)
     );
-  }
-  /**
-   * Sets the new social contract ID.
-   * @param {string} contractId - the account of the new social contract ID.
-   */
-  public setContractId(contractId: string): void {
-    this.contractId = contractId;
   }
 }