+++ src/types/NetworkIds.ts
@@ -1,0 +1,3 @@
+type NetworkIds = 'testnet' | 'mainnet' | 'betanet' | 'localnet';
+
+export default NetworkIds;