+++ src/utils/rpcQueries/viewAccessKeyList/types/IOptions.ts
@@ -1,0 +1,8 @@
+import { providers } from 'near-api-js';
+
+interface IOptions {
+  accountID: string;
+  provider: providers.JsonRpcProvider;
+}
+
+export default IOptions;